less uint16

+++ tests/core/data_layer/test_data_rpc.py
@@ -147,7 +147,7 @@ async def test_create_double_insert(one_wallet_node_and_rpc: nodes) -> None:
     )
     await time_out_assert(15, wallet_node.wallet_state_manager.main_wallet.get_confirmed_balance, funds)
     wallet_rpc_api = WalletRpcApi(wallet_node)
-    async for data_layer in init_data_layer(root_path, wallet_rpc_port=uint16(wallet_node.server._port)):
+    async for data_layer in init_data_layer(root_path, wallet_rpc_port=wallet_node.server._port):
         data_rpc_api = DataLayerRpcApi(data_layer)
         res = await data_rpc_api.create_data_store({})
         assert res is not None
@@ -210,7 +210,7 @@ async def test_keys_values_ancestors(one_wallet_node_and_rpc: nodes) -> None:
     await time_out_assert(15, wallet_node.wallet_state_manager.main_wallet.get_confirmed_balance, funds)
     wallet_rpc_api = WalletRpcApi(wallet_node)
     # TODO: with this being a pseudo context manager'ish thing it doesn't actually handle shutdown
-    async for data_layer in init_data_layer(root_path, wallet_rpc_port=uint16(wallet_node.server._port)):
+    async for data_layer in init_data_layer(root_path, wallet_rpc_port=wallet_node.server._port):
         data_rpc_api = DataLayerRpcApi(data_layer)
         res = await data_rpc_api.create_data_store({})
         assert res is not None
@@ -294,7 +294,7 @@ async def test_get_roots(one_wallet_node_and_rpc: nodes) -> None:
     )
     await time_out_assert(15, wallet_node.wallet_state_manager.main_wallet.get_confirmed_balance, funds)
     wallet_rpc_api = WalletRpcApi(wallet_node)
-    async for data_layer in init_data_layer(root_path, wallet_rpc_port=uint16(wallet_node.server._port)):
+    async for data_layer in init_data_layer(root_path, wallet_rpc_port=wallet_node.server._port):
         data_rpc_api = DataLayerRpcApi(data_layer)
         res = await data_rpc_api.create_data_store({})
         assert res is not None
@@ -369,7 +369,7 @@ async def test_get_root_history(one_wallet_node_and_rpc: nodes) -> None:
     )
     await time_out_assert(15, wallet_node.wallet_state_manager.main_wallet.get_confirmed_balance, funds)
     wallet_rpc_api = WalletRpcApi(wallet_node)
-    async for data_layer in init_data_layer(root_path, wallet_rpc_port=uint16(wallet_node.server._port)):
+    async for data_layer in init_data_layer(root_path, wallet_rpc_port=wallet_node.server._port):
         data_rpc_api = DataLayerRpcApi(data_layer)
         res = await data_rpc_api.create_data_store({})
         assert res is not None
@@ -447,7 +447,7 @@ async def test_get_kv_diff(one_wallet_node_and_rpc: nodes) -> None:
     )
     await time_out_assert(15, wallet_node.wallet_state_manager.main_wallet.get_confirmed_balance, funds)
     wallet_rpc_api = WalletRpcApi(wallet_node)
-    async for data_layer in init_data_layer(root_path, wallet_rpc_port=uint16(wallet_node.server._port)):
+    async for data_layer in init_data_layer(root_path, wallet_rpc_port=wallet_node.server._port):
         data_rpc_api = DataLayerRpcApi(data_layer)
         res = await data_rpc_api.create_data_store({})
         assert res is not None
@@ -538,7 +538,7 @@ async def test_batch_update_matches_single_operations(one_wallet_node_and_rpc: n
     )
     await time_out_assert(15, wallet_node.wallet_state_manager.main_wallet.get_confirmed_balance, funds)
     wallet_rpc_api = WalletRpcApi(wallet_node)
-    async for data_layer in init_data_layer(root_path, wallet_rpc_port=uint16(wallet_node.server._port)):
+    async for data_layer in init_data_layer(root_path, wallet_rpc_port=wallet_node.server._port):
         data_rpc_api = DataLayerRpcApi(data_layer)
         res = await data_rpc_api.create_data_store({})
         assert res is not None
